Cleanup, move to extension, and add conditional compilation
